% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_batting.R
\name{aim.batting.wrapper}
\alias{aim.batting.wrapper}
\title{Wrapper function for cv.aim.batting to be passed to kfold.cv.}
\usage{
aim.batting.wrapper(data, args)
}
\arguments{
\item{data}{data frame equal to cbind(y, x), where y and x are inputs to aim.batting.}

\item{args}{list containing all other input arguments to aim.batting except for x and y.}
}
\value{
prediction rule as returned by aim.batting.
}
\description{
Wrapper function for cv.aim.batting to be passed to kfold.cv.
}
