% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_batting.R
\name{aim.rule.batting.wrapper}
\alias{aim.rule.batting.wrapper}
\title{Wrapper function for aim.rule.batting, to be passed to kfold.cv}
\usage{
aim.rule.batting.wrapper(data, args)
}
\arguments{
\item{data}{data frame equal to cbind(y, x), where y and x are inputs to aim.rule.batting.}

\item{args}{list containing all other input arguments to aim.rule.batting except for x and y.}
}
\value{
prediction rule returned by aim.rule.batting.
}
\description{
Wrapper function for aim.rule.batting, to be passed to kfold.cv
}
