% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_batting.R
\name{aim.score.pred}
\alias{aim.score.pred}
\title{Find score of cutoff (returned by aim.find.cutoff.pred) for predictive case.}
\usage{
aim.score.pred(data, yvar, censorvar, trtvar, trtref, xvar, type, cutoff, nsubj,
  min.sigp.prcnt)
}
\arguments{
\item{data}{data frame containing the response, covariate, treatment variable and censoring variable (only for time to event response).}

\item{yvar}{response variable name.}

\item{censorvar}{censoring variable name 1:event; 0: censor.}

\item{trtvar}{treatment variable name.}

\item{trtref}{code for treatment arm.}

\item{xvar}{covariate variable name.}

\item{type}{"c" continuous; "s" survival; "b" binary.}

\item{cutoff}{cutpoint of interest.}

\item{nsubj}{number of subjects.}

\item{min.sigp.prcnt}{desired proportion of signature positive group size.}
}
\value{
AIM score for a single covariate-cutoff combination.
}
\description{
Find score of cutoff (returned by aim.find.cutoff.pred) for predictive case.
}
