% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_batting.R
\name{backfit.cox.interaction}
\alias{backfit.cox.interaction}
\title{An internal function used in cox.interaction}
\usage{
backfit.cox.interaction(x, trt, y, delta, cutp, mincut = 0)
}
\arguments{
\item{x}{the predictor matrix.}

\item{trt}{the treatment indicator vector.}

\item{y}{the vector of the time to event response variable.}

\item{delta}{status indicator: 1=failure 0=alive}

\item{cutp}{a specific cutpoint}

\item{mincut}{the minimum cutting proportion for the binary rule at either end. It typically is between 0 and 0.2. It is the parameter in the functions of AIM package.}
}
\description{
An internal function used in cox.interaction
}
