% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{balanced.folds}
\alias{balanced.folds}
\title{Create balanced folds for cross-validation.}
\usage{
balanced.folds(y, nfolds = min(min(table(y)), 10))
}
\arguments{
\item{y}{the response vector}

\item{nfolds}{number of folds}
}
\value{
This function returns balanced folds
}
\description{
Create balanced folds for cross-validation.
}
