% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_batting.R
\name{batting.prog}
\alias{batting.prog}
\title{Main prognostic BATTing function}
\usage{
batting.prog(dataset, ids, yvar, censorvar, type, class.wt, xvar, n.boot,
  des.res, min.sigp.prcnt)
}
\arguments{
\item{dataset}{input dataset in data frame}

\item{ids}{training indices}

\item{yvar}{response variable name}

\item{censorvar}{censoring variable name 1:event; 0: censor.}

\item{type}{"c" continuous; "s" survival; "b" binary}

\item{class.wt}{vector of length 2 used to weight the accuracy score , useful when there is class imbalance in binary data defaults to c(1,1)}

\item{xvar}{name of predictor for which cutpoint needs to be obtained}

\item{n.boot}{number of bootstraps for BATTing step.}

\item{des.res}{the desired response. "larger": prefer larger response. "smaller": prefer smaller response.}

\item{min.sigp.prcnt}{desired proportion of signature positive group size for a given cutoff.}
}
\value{
a signature rule consisting of variable name, direction, optimal cutpoint and the corresponding p-value.
}
\description{
Main prognostic BATTing function
}
