% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{binary.stats}
\alias{binary.stats}
\title{A function for binary statistics}
\usage{
binary.stats(pred.class, y.vec)
}
\arguments{
\item{pred.class}{predicted output for each subject}

\item{y.vec}{response vector}
}
\value{
a data frame with sensitivity, specificity, NPV, PPV and accuracy
}
\description{
A function for binary statistics
}
