% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{combine.condition}
\alias{combine.condition}
\title{Internal function}
\usage{
combine.condition(trace.inside.condition, new.inside.condition)
}
\arguments{
\item{trace.inside.condition}{list of signature rules}

\item{new.inside.condition}{new signature rule}
}
\value{
updated list of signature rules
}
\description{
Internal function
}
