% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{create.training.dataset.index}
\alias{create.training.dataset.index}
\title{create training/testing dataset indexes.}
\usage{
create.training.dataset.index(training.percent, n)
}
\arguments{
\item{training.percent}{percentage of subjects in training data as mentioned in prim.train function.}

\item{n}{number of sbjects in the whole dataset.}
}
\value{
a list containing training and test data indices.
}
\description{
create training/testing dataset indexes.
}
