% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{cv.folds}
\alias{cv.folds}
\title{Cross-validation folds.}
\usage{
cv.folds(n, folds = 10)
}
\arguments{
\item{n}{number of observations.}

\item{folds}{number of folds.}
}
\value{
a list containing the observation numbers for each fold.
}
\description{
Cross-validation folds.
}
