% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{cv.pval}
\alias{cv.pval}
\title{p-value calculation for each iteration of cross validation.}
\usage{
cv.pval(yvar, censorvar = NULL, trtvar = NULL, data, type = "s")
}
\arguments{
\item{yvar}{response variable name.}

\item{censorvar}{censor-variable name.}

\item{trtvar}{treatment variable name. For prognostic case trtvar=NULL.}

\item{data}{dataset containing response and predicted output.}

\item{type}{data type - "c" - continuous , "b" - binary, "s" - time to event  - default = "c".}
}
\value{
p-value based on response and prediction vector for each iteration.
}
\description{
p-value calculation for each iteration of cross validation.
}
