% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{evaluate.cv.results}
\alias{evaluate.cv.results}
\title{Cross-validation Performance Evaluation}
\usage{
evaluate.cv.results(cv.data, y, censor.vec, trt.vec, type)
}
\arguments{
\item{cv.data}{output of prediction function from kfold.cv}

\item{y}{data frame of the response variable from CV data.}

\item{censor.vec}{data frame indicating censoring for survival data. For binary or continuous data, set censor.vec <- NULL.}

\item{trt.vec}{data frame indicating whether or not the patient was treated. For the pronostic case, set trt.vec <- NULL.}

\item{type}{data type - "c" - continuous , "b" - binary, "s" - time to event  - default = "c"}
}
\value{
a list containing raw statistics and fold information
}
\description{
Take the raw output of kfold.cv and calculate performance statistics for each iteration of the cross-validation.
}
