% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{evaluate.results}
\alias{evaluate.results}
\title{Get statistics for a single set of predictions.}
\usage{
evaluate.results(y, predict.data, censor.vec = NULL, trt.vec = NULL,
  trtref = NULL, type)
}
\arguments{
\item{y}{data frame of the response variable.}

\item{predict.data}{output of prediction function from kfold.cv.}

\item{censor.vec}{data frame indicating censoring for survival data. For binary or continuous data, set censor.vec <- NULL.}

\item{trt.vec}{data frame indicating whether or not the patient was treated. For the pronostic case, set trt.vec <- NULL.}

\item{trtref}{treatment reference.}

\item{type}{data type - "c" - continuous , "b" - binary, "s" - time to event  - default = "c".}
}
\value{
a list containing p-value and group statistics.
}
\description{
Get statistics for a single set of predictions.
}
