% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.glmnet}
\alias{filter.glmnet}
\title{Flitering using MC glmnet}
\usage{
filter.glmnet(data, type, yvar, xvars, censorvar, trtvar, trtref, n.boot = 50,
  cv.iter = 20, pre.filter = length(xvars))
}
\arguments{
\item{data}{input data frame}

\item{type}{"c" continuous; "s" survival; "b" binary}

\item{yvar}{response variable name}

\item{xvars}{covariates variable name}

\item{censorvar}{censoring variable name 1:event; 0: censor.}

\item{trtvar}{treatment variable name}

\item{trtref}{code for treatment arm}

\item{n.boot}{number of bootstrap for filtering}

\item{cv.iter}{number of iterations required for MC glmnet filtering}

\item{pre.filter}{NULL, no prefiltering conducted;"opt", optimized number of predictors selected; An integer: min(opt, integer) of predictors selected}
}
\value{
variables selected after glmnet filtering
}
\description{
Flitering using MC glmnet
}
