% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.univariate}
\alias{filter.univariate}
\title{Univariate Filtering}
\usage{
filter.univariate(data, type, yvar, xvars, censorvar, trtvar, trtref = 1,
  pre.filter = length(xvars))
}
\arguments{
\item{data}{input data frame}

\item{type}{"c" continuous; "s" survival; "b" binary}

\item{yvar}{response variable name}

\item{xvars}{covariates variable name}

\item{censorvar}{censoring variable name 1:event; 0: censor.}

\item{trtvar}{treatment variable name}

\item{trtref}{code for treatment arm}

\item{pre.filter}{NULL, no prefiltering conducted;"opt", optimized number of predictors selected; An integer: min(opt, integer) of predictors selected}
}
\value{
covariate names after univariate filtering.
}
\description{
Univariate Filtering
}
