% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_batting.R
\name{get.var.counts.aim}
\alias{get.var.counts.aim}
\title{Get counts of signature variables from output of cv.aim.batting.}
\usage{
get.var.counts.aim(sig.list, xvars)
}
\arguments{
\item{sig.list}{signature list output from cv.aim.batting.}

\item{xvars}{predictor variable names.}
}
\value{
counts of signature variables.
}
\description{
Get counts of signature variables from output of cv.aim.batting.
}
