% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_batting.R
\name{get.var.counts.seq}
\alias{get.var.counts.seq}
\title{Get signature variables from output of seqlr.batting.}
\usage{
get.var.counts.seq(sig.list, xvars)
}
\arguments{
\item{sig.list}{signature list returned by seqlr.batting.}

\item{xvars}{predictor variable names}
}
\value{
the variables included in signature rules returned by seqlr.batting
}
\description{
Get signature variables from output of seqlr.batting.
}
