% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{interaction.plot}
\alias{interaction.plot}
\title{A function for interaction plot}
\usage{
interaction.plot(data.eval, type, main = "Interaction Plot",
  trt.lab = c("Trt.", "Ctrl."))
}
\arguments{
\item{data.eval}{output of evaluate.results or summarize.cv.stats}

\item{type}{data type - "c" - continuous , "b" - binary, "s" - time to event  - default = "c".}

\item{main}{title of the plot}

\item{trt.lab}{treatment label}
}
\value{
A ggplot object.
}
\description{
A function for interaction plot
}
