% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{make.arg.list}
\alias{make.arg.list}
\title{Create a list of variables corresponding to the arguments of the function func.name and assigns values.}
\usage{
make.arg.list(func.name)
}
\arguments{
\item{func.name}{function name}
}
\value{
list of variables corresponding to the arguments of the function
}
\description{
Create a list of variables corresponding to the arguments of the function func.name and assigns values.
}
