% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{permute.vector}
\alias{permute.vector}
\title{Randomly permute the entries of a vector.}
\usage{
permute.vector(x)
}
\arguments{
\item{x}{the vector for which its entries have to be permuted}
}
\value{
the permuted vector
}
\description{
Randomly permute the entries of a vector.
}
