% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_batting.R
\name{pred.aim.cv}
\alias{pred.aim.cv}
\title{Make predictions for data given prediction rule in predict.rule}
\usage{
pred.aim.cv(data, predict.rule, args)
}
\arguments{
\item{data}{Data frame of form cbind(y, x), where y and x are inputs to cv.seq.batting.}

\item{predict.rule}{Prediction rule returned by seq.batting.}

\item{args}{list of the form list(xvar=xvar, yvar=yvar)}
}
\value{
The input data with an added column, a logical vector indicating the prediction for each row of data.
}
\description{
Make predictions for data given prediction rule in predict.rule
}
