% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aim_batting.R
\name{pred.aim}
\alias{pred.aim}
\title{Make predictions for data given prediction rule in predict.rule.}
\usage{
pred.aim(x, predict.rule)
}
\arguments{
\item{x}{the predictor matrix.}

\item{predict.rule}{prediction rule returned by seq.batting.}
}
\value{
The input data with an added column, a logical vector indicating the prediction for each row of data.
}
\description{
Make predictions for data given prediction rule in predict.rule.
}
