% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{pred.prim.cv}
\alias{pred.prim.cv}
\title{Prediction function for PRIM CV}
\usage{
pred.prim.cv(data, predict.rule, args)
}
\arguments{
\item{data}{input data frame}

\item{predict.rule}{signature rules as returned by prim.train}

\item{args}{list of the form list(yvar=yvar)}
}
\value{
The input data with an added column, a logical vector indicating the prediction for each row of data.
}
\description{
Prediction function for PRIM CV
}
