% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{pred.prim}
\alias{pred.prim}
\title{Prediction function for PRIM}
\usage{
pred.prim(data, predict.rule)
}
\arguments{
\item{data}{input data frame (only covariates)}

\item{predict.rule}{signature rules returned by prim.train}
}
\value{
The input data with an added column, a logical vector indicating the prediction for each row of data.
}
\description{
Prediction function for PRIM
}
