% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_batting.R
\name{pred.seqlr.cv}
\alias{pred.seqlr.cv}
\title{Prediction function for CV Sequential BATTing}
\usage{
pred.seqlr.cv(data, predict.rule, args)
}
\arguments{
\item{data}{input data frame}

\item{predict.rule}{Prediction rule returned by seqlr.batting.}

\item{args}{Prediction rule arguments}
}
\value{
a logical vector indicating the prediction for each row of data.
}
\description{
Assign positive and negative groups for cross-validation data given prediction rule in predict.rule.
}
