% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_batting.R
\name{pred.seqlr}
\alias{pred.seqlr}
\title{Prediction function for Sequential BATTing}
\usage{
pred.seqlr(x, predict.rule)
}
\arguments{
\item{x}{input predictors matrix}

\item{predict.rule}{Prediction rule returned by seqlr.batting.}
}
\value{
a logical vector indicating the prediction for each row of data.
}
\description{
Assign positive and negative groups based on predict.rule, the output of seqlr.batting.
}
