% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{prim.train.wrapper}
\alias{prim.train.wrapper}
\title{Wrapper function for PRIM CV}
\usage{
prim.train.wrapper(data, args)
}
\arguments{
\item{data}{input data frame}

\item{args}{list containing all other input arguments to prim.train except for x and y.}
}
\value{
prediction rule as returned by prim.train
}
\description{
Wrapper function for PRIM CV
}
