% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{pval.cal.prog}
\alias{pval.cal.prog}
\title{Calculate p-value for treatment in each subgroup in prognostic case}
\usage{
pval.cal.prog(data, yvar, censorvar, grp.id, type, des.res)
}
\arguments{
\item{data}{input data frame}

\item{yvar}{name for response variable}

\item{censorvar}{name for censoring (1: event; 0: censor), default = NULL}

\item{grp.id}{subgroup id}

\item{type}{type of response variable: "c" continuous (default); "s" survival; "b" binary}

\item{des.res}{the desired response. "larger": prefer larger response (default) "smaller": prefer smaller response}
}
\value{
p-value for the main effect given the dataset
}
\description{
Calculate p-value for treatment in each subgroup in prognostic case
}
