% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{pval.cal}
\alias{pval.cal}
\title{Calculate p-value for treatment in each subgroup in predictive case}
\usage{
pval.cal(data, yvar, censorvar, trtvar, type, des.res)
}
\arguments{
\item{data}{input data frame}

\item{yvar}{name for response variable}

\item{censorvar}{name for censoring (1: event; 0: censor), default = NULL}

\item{trtvar}{name for treatment variable, default = NULL (prognostic signature)}

\item{type}{type of response variable: "c" continuous (default); "s" survival; "b" binary}

\item{des.res}{the desired response. "larger": prefer larger response (default) "smaller": prefer smaller response}
}
\value{
p-value for the treatment given the dataset
}
\description{
Calculate p-value for treatment in each subgroup in predictive case
}
