% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_batting.R
\name{query.data}
\alias{query.data}
\title{internal function used in seqlr.batting}
\usage{
query.data(data, rule)
}
\arguments{
\item{data}{the given dataset}

\item{rule}{rule is a vector of the form [x-variable, direction, cutoff, p-value]}
}
\value{
a logical variable indicating whether rules are satisfied or not.
}
\description{
internal function used in seqlr.batting
}
