% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prim.R
\name{query.from.condition}
\alias{query.from.condition}
\title{An internal function inside one.pasting.}
\usage{
query.from.condition(d, condition, g.str = ">=", l.str = "<=")
}
\arguments{
\item{d}{dataset for subjects in consideration.}

\item{condition}{signature rule in consideration.}

\item{g.str}{">="}

\item{l.str}{"<="}
}
\value{
a vector of logical arguments indicating whether the conditions can be  satisfied for the subjects in d.
}
\description{
An internal function inside one.pasting.
}
