% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{resample}
\alias{resample}
\title{Creates a permutation of given size.}
\usage{
resample(x, size, ...)
}
\arguments{
\item{x}{the x vector.}

\item{size}{resampling size.}

\item{...}{optional argument.}
}
\value{
A resample of x is returned.
}
\description{
Creates a permutation of given size.
}
