\name{Sepsis.train}
\docType{data}
\alias{Sepsis.train}
\title{Sepsis Trial training dataset}
\description{
  This is a simulated dataset based on a Phase III clinical trial compared a novel treatment to the standard of care (control) in patients with severe sepsis. The outcome of interest is a binary endpoint indicating subjects’ death after 28 days of treatment. Available markers include demographic and clinical covariates, i.e., age, time from first sepsis-organ fail to start drug, sum of baseline SOFA socres (cardiovascular, hematology, hepaticrenal, and respiration scores), number of baseline organ failures, pre-infusion apache-ii score, baseline GLASGOW coma scale score, baseline activity of daily living score; and laboratory markers, i.e., baseline local platelets, creatinine, serum IL-6 concentration, local bilirunbin. 
}
\usage{Sepsis.train}
\format{Dataset as a data frame}
\source{http://multxpert.com}
\references{
   Lipkovich I, Dmitrienko A, Denne J, Enas G (2011) Subgroup identification based on differential effect search--a recursive partitioning method for establishing response to treatment in patient subpopulations. Stat Med 30:2601–2621. doi: 10.1002/sim.4289
}
