% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_batting.R
\name{seqlr.batting.wrapper}
\alias{seqlr.batting.wrapper}
\title{Wrapper function for seqlr.batting, to be passed to kfold.cv.}
\usage{
seqlr.batting.wrapper(data, args)
}
\arguments{
\item{data}{data frame equal to cbind(y, x, trt, censor), where y and x are inputs to seqlr.batting.}

\item{args}{list containing all other input arguments to seq.batting except for x and y. Also contains xvars=names(x) and yvar=names(y).}
}
\value{
prediction rule returned by seqlr.batting.
}
\description{
Wrapper function for seqlr.batting, to be passed to kfold.cv.
}
