% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_batting.R
\name{seqlr.find.cutoff.pred}
\alias{seqlr.find.cutoff.pred}
\title{Find cutoff for predictive case.}
\usage{
seqlr.find.cutoff.pred(data, yvar, censorvar, xvar, trtvar, type, class.wt, dir,
  nsubj, min.sigp.prcnt)
}
\arguments{
\item{data}{input data frame.}

\item{yvar}{response variable name.}

\item{censorvar}{censoring variable name.}

\item{xvar}{name of predictor for which cutpoint needs to be obtained.}

\item{trtvar}{treatment variable name.}

\item{type}{"c" continuous; "s" survival; "b" binary.}

\item{class.wt}{vector of length 2 used to weight the accuracy score , useful when there is class imbalance in binary data defaults to c(1,1).}

\item{dir}{direction of cut.}

\item{nsubj}{number of subjects.}

\item{min.sigp.prcnt}{desired proportion of signature positive group size for a given cutoff.}
}
\value{
the optimal score (p-value of subgroup*treatment interaction) for a predictor variable.
}
\description{
Find cutoff for predictive case.
}
