% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_functions.R
\name{summarize.cv.stats}
\alias{summarize.cv.stats}
\title{Calculate summary statistics from raw statistics returned by evaluate.cv.results.}
\usage{
summarize.cv.stats(raw.stats, trtvar, type)
}
\arguments{
\item{raw.stats}{raw statistics from evaluate.cv.results}

\item{trtvar}{treatment variable name}

\item{type}{data type - "c" - continuous , "b" - binary, "s" - time to event  - default = "c"}
}
\value{
a list containing p-values, summary statistics and group statistics.
}
\description{
Calculate summary statistics from raw statistics returned by evaluate.cv.results.
}
